# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FitManager API
  version: 1.0.0
  description: Contrato HTTP de FitManager. Todos los recursos autenticados se aíslan por gimnasio desde el contexto JWT.
servers:
  - url: http://localhost:3000/api
    description: Desarrollo local
  - url: https://fitmanager-backend-nine.vercel.app/api
    description: Producción
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginStaff
      security: []
      requestBody: {$ref: '#/components/requestBodies/Login'}
      responses: {'200': {$ref: '#/components/responses/Auth'}, '401': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/Error'}}
  /auth/login-cliente:
    post:
      tags: [Auth]
      operationId: loginCliente
      security: []
      requestBody: {$ref: '#/components/requestBodies/Login'}
      responses: {'200': {$ref: '#/components/responses/Auth'}, '401': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/Error'}}
...
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      security: [{cookieAuth: []}]
      parameters: [{$ref: '#/components/parameters/CsrfToken'}]
      responses: {'200': {$ref: '#/components/responses/Auth'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}
...
  /auth/forgot-password:
    post: {tags: [Auth], operationId: forgotPassword, security: [], responses: {'200': {$ref: '#/components/responses/Ok'}, '429': {$ref: '#/components/responses/Error'}}}
  /auth/reset-password:
    post: {tags: [Auth], operationId: resetPassword, security: [], responses: {'200': {$ref: '#/components/responses/Ok'}, '400': {$ref: '#/components/responses/Error'}}}
...
  /auth/setup-password:
    post: {tags: [Auth], operationId: setupPassword, security: [], responses: {'200': {$ref: '#/components/responses/Ok'}, '400': {$ref: '#/components/responses/Error'}}}
  /gimnasios:
    post: {tags: [Gimnasios], operationId: registerGym, security: [], responses: {'201': {$ref: '#/components/responses/Auth'}, '409': {$ref: '#/components/responses/Error'}}}
  /usuarios:
    get: {tags: [Usuarios], operationId: listUsers, responses: {'200': {$ref: '#/components/responses/List'}, '403': {$ref: '#/components/responses/Error'}}}
    post: {tags: [Usuarios], operationId: createUser, responses: {'201': {$ref: '#/components/responses/Created'}, '409': {$ref: '#/components/responses/Error'}}}
...
  /clientes:
    get: {tags: [Clientes], operationId: listClients, responses: {'200': {$ref: '#/components/responses/List'}}}
    post: {tags: [Clientes], operationId: createClient, responses: {'201': {$ref: '#/components/responses/Created'}, '409': {$ref: '#/components/responses/Error'}}}
...
  /clientes-membresias:
    get: {tags: [Membresías], operationId: listMemberships, responses: {'200': {$ref: '#/components/responses/List'}}}
    post: {tags: [Membresías], operationId: assignMembership, responses: {'201': {$ref: '#/components/responses/Created'}, '409': {$ref: '#/components/responses/Error'}}}
...
  /transferencias:
    get: {tags: [Transferencias], operationId: listTransfers, responses: {'200': {$ref: '#/components/responses/List'}}}
    post: {tags: [Transferencias], operationId: createTransfer, responses: {'201': {$ref: '#/components/responses/Created'}, '409': {$ref: '#/components/responses/Error'}}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: fitmanager_refresh
  parameters:
...
    CsrfToken: {name: X-CSRF-Token, in: header, required: true, schema: {type: string, minLength: 32}}
  requestBodies:
    Login:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [correo, password]
            properties:
              correo: {type: string, format: email}
              password: {type: string, format: password}
  responses:
    Ok: {description: Operación exitosa, content: {application/json: {schema: {type: object, additionalProperties: true}}}}
    Created: {description: Recurso creado, content: {application/json: {schema: {type: object, additionalProperties: true}}}}
    List: {description: Lista tenant-safe, content: {application/json: {schema: {type: array, items: {type: object, additionalProperties: true}}}}}
    Auth: {description: Sesión creada o renovada, content: {application/json: {schema: {$ref: '#/components/schemas/AuthResponse'}}}}
...
    Error: {description: Error sin detalles sensibles, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  schemas:
...
    AuthResponse:
      type: object
      required: [token, csrfToken]
      properties:
        token: {type: string}
        csrfToken: {type: string}
        usuario: {type: object, additionalProperties: true}
        cliente: {type: object, additionalProperties: true}
    Error:
      type: object
      required: [error]
      properties:
        error: {type: string}
        codigo: {type: string}
        data: {type: object, additionalProperties: true}